Pause, Menu Enhancement
Added a sort-of pause to the game, so it at least stops when you press
start. actual graphics for that needed, coming later.
Reworked some of the menus to now work with fire & jump buttons instead
of only the start button.
gameover-scene now takes you back to the game automatically after 5
seconds